added searchbar + btn + added api with results
added a search bar and button, connected api and made it funtional with search. also added a popup page that shows book results in a still position that has a favorite button on them. also made a category page that lets you press a category and then it shows you book alternatives in the pop up + carousel. (aslo changed some minor css)